add clean_artifacts.ts to run chown and remove files
