+++ package.json
@@ -27,7 +27,6 @@
   "dependencies": {
     "@actions/core": "^1.3.0",
     "@actions/exec": "^1.0.4",
-    "@actions/glob": "^0.1.2",
     "@actions/io": "^1.1.0",
     "js-yaml": "^4.1.0",
     "node-fetch": "^2.6.1"
