+++ src/ort.ts
@@ -1,4 +1,3 @@
-import {userInfo} from 'os'
 import {run_docker_container, volume2dockerargs} from './docker'
 
 /**
@@ -60,21 +59,4 @@ export async function report(output_dir: string): Promise<void> {
         '/out'
     ]
     await run_docker_container(docker_args, ort_args)
-}
-
-export async function chown(
-    input_dir: string,
-    output_dir: string
-): Promise<void> {
-    const volumes = {
-        [input_dir]: '/in',
-        [output_dir]: '/out'
-    }
-    const docker_args: string[] = volume2dockerargs(volumes)
-
-    const {uid, gid} = userInfo()
-
-    docker_args.push('--entrypoint', 'chown')
-    const chown_args = ['-R', `${uid}:${gid}`, '/in', '/out']
-    await run_docker_container(docker_args, chown_args)
 }
